moved version modifier and separator enum to top level
